Refactor and fix.
* Add UNLICENSE
* Remove Pipfile.lock
* Use argparse instead of click
* Switch to python3
