+++ requirements-dev.txt
@@ -1,19 +1,4 @@
-blessings==1.6
 bpython==0.16
-certifi==2017.7.27.1
-chardet==3.0.4
-click==6.7
-coverage
-curtsies==0.2.11
-greenlet==0.4.12
-idna==2.6
-pkg-release
 PlexAPI==3.0.3
 pre-commit>=0.15.0
-Pygments==2.2.0
 pytest
-requests==2.18.4
-six==1.11.0
-urllib3==1.22
-wcwidth==0.1.7
-websocket-client==0.44.0
