+++ requirements.txt
@@ -1,15 +1,1 @@
-blessings==1.6
-bpython==0.16
-certifi==2017.7.27.1
-chardet==3.0.4
-click==6.7
-curtsies==0.2.11
-greenlet==0.4.12
-idna==2.6
-PlexAPI==3.0.3
-Pygments==2.2.0
-requests==2.18.4
-six==1.11.0
-urllib3==1.22
-wcwidth==0.1.7
-websocket-client==0.44.0
+PlexAPI>=3.0.3
